fix: hooks call order
