+++ libs/frontend/modules/builder/src/renderer/renderContainer/RenderContainer.tsx
@@ -1,54 +1,26 @@
-import { DATA_ID } from '@codelab/frontend/abstract/core'
-import { PropsDataByElementId } from '@codelab/shared/abstract/core'
-import { Maybe, Nullable } from '@codelab/shared/abstract/types'
-import { isArray, merge } from 'lodash'
-import React, { ReactElement, useCallback, useEffect } from 'react'
+import React, { ReactElement, useEffect } from 'react'
+import { getRenderedProps } from '../utils'
 import { RenderContainerProps } from './types'
 
-export const RenderContainer = ({
-  context,
-  children: rendered,
-  isRoot = false,
-}: RenderContainerProps) => {
-  const getElementId = (renderedElement: ReactElement): Maybe<string> =>
-    renderedElement?.props?.[DATA_ID]
+export const RenderContainer = ({ context, root }: RenderContainerProps) => {
+  if (context?.inspect) {
+    console.group('Root')
+  }
 
-  const treeToObject = useCallback(
-    (renderElement: ReactElement): Nullable<PropsDataByElementId> => {
-      if (!renderElement) {
-        return null
-      }
+  const rendered = context?.render(root, context, {})
 
-      if (isArray(renderElement)) {
-        return renderElement.map(treeToObject).reduce(merge)
-      }
-
-      const { children } = renderElement?.props || {}
-
-      const childrenQueue =
-        children && isArray(children) ? children : [children]
-
-      const childrenProps = childrenQueue
-        .filter((x) => !!x)
-        .map(treeToObject)
-        .reduce(merge, {})
-
-      const elementId = getElementId(renderElement)
-      const elementProps = elementId ? { [elementId]: renderElement.props } : {}
-
-      return { ...elementProps, ...childrenProps }
-    },
-    [],
-  )
+  if (context?.inspect) {
+    console.groupEnd()
+  }
 
   useEffect(() => {
-    if (!context?.onRendered || !isRoot) {
+    if (!context?.onRendered) {
       return
     }
 
-    const renderMap = treeToObject(rendered as ReactElement)
+    const renderMap = getRenderedProps(rendered as ReactElement)
     context.onRendered(renderMap || {})
-  }, [context, rendered, isRoot, treeToObject])
+  }, [rendered, context])
 
   return <>{rendered}</>
 }
