+++ libs/frontend/modules/builder/src/renderer/utils/index.ts
@@ -2,6 +2,7 @@ export * from './applyBinding'
 export * from './containerKey'
 export * from './evalCss'
 export * from './getPropsByTypeKind'
+export * from './getRenderedProps'
 export * from './logRendered'
 export * from './queryRenderedElementById'
 export * from './shouldStopRendering'
